+++ docs/start/changelog.rst
@@ -9,6 +9,8 @@ ChangeLog
       For example::
 
         $ lifx 'lan(default_broadcast="10.1.1.255"):get_attr' _ color
+   * Added a default ``chain`` plan for the Gatherer to use for getting tile
+     chain information
 
 0.24.3 - 18 December 2019
     * Fixed a bug where response packets were matched to the wrong requests